medrequirements.txt
datasets==2.19.1
huggingface_hub==0.24.1
pandas==2.2.2
peft==0.11.1
scikit_learn==1.4.2
torch==2.3.1
transformers==4.43.1
trl==0.9.6
bitsandbytes==0.43.3
tensorboardX
chromadb
openpyxl==3.1.2
PyYAML==6.0.1